New Navigation Service
I've implemented a whole new navigation service architecture that will simplify a lot the development.
A developer can now use these functions very easily to the navigation between page :
- NavigateTo()
- GoBack()

There is also an event called "NavigationChanged" to allow windows and page to subscribe easily to the navigation service and be aware when a page has been changed. There is 2 location in the code where this event is used:
-For the MainWindows to know when to point to a new page
- For the SimplePageViewModel, to know when the page has been changed, like this it can reset his property Name

More will come to enchance this new navigaiton Service, but now I consider it as fully fonctionnal and useful to make it easier to code now